OPEN - issue CL-8742: [Server] Update CreatePendingPaymentJob to process o2 customers. Implementation of a new payment system http://jira.dev.now-technologies.mobi:8181/browse/CL-8742
bug fixed

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserRepository.java
@@ -149,7 +149,7 @@ public interface UserRepository extends PagingAndSortingRepository<User, Integer
 			"or u.lastSubscribedPaymentSystem is null) " +
 			"and u.deactivatedGraceCreditMillis=0 " +
 			"and u.nextSubPayment<?1))")
-	List<User> getListOfUsersForWeeklyUpdate(int epochSeconds, Pageable pageable);
+	List<User> getListOfUsersForWeeklyUpdate(long epochSeconds, Pageable pageable);
 
 	@Modifying
 	@Query(value="update User user " +
